requirements.txt
python-dotenv
google-genai
chromadb
fastapi
uvicorn
